# src/main/java/com/hrs/request/UpdateCheckInRequest.java
package com.hrs.request;

import lombok.AccessLevel;
import lombok.Data;
import lombok.experimental.FieldDefaults;

@Data
@FieldDefaults(level = AccessLevel.PRIVATE)
public class UpdateCheckInRequest {
  boolean checkedIn;
}