# src/main/java/com/hrs/request/UpdateParcelAvailableRequest.java
package com.hrs.request;

import lombok.AccessLevel;
import lombok.Data;
import lombok.experimental.FieldDefaults;

@Data
@FieldDefaults(level = AccessLevel.PRIVATE)
public class UpdateParcelAvailableRequest {
  boolean parcelAvailable;
}